feat: support join-group api request
