+++ src/services/mod.rs
@@ -1,1 +1,2 @@
+pub mod group;
 pub mod user;
